Restructure demo
I restructured the project, so now it has a demo folder which includes my testing files such as html, css and a default script file.
I also exported the main class.
