+++ demo/index.html
@@ -10,6 +10,6 @@
       <div class="content">asd</div>
       <div class="content">asd</div>
     </div>
-    <script src="./src/custom-scrollbar.js" type="module"></script>
+    <script src="./script.js" type="module"></script>
   </body>
 </html>
